+++ src/Agents.jl
@@ -28,6 +28,7 @@ include("simulations/sample.jl")
 # 4.0 Depreciations
 @deprecate space_neighbors nearby_agents
 @deprecate node_neighbors nearby_positions
+@deprecate get_node_contents agents_in_pos
 
 # Predefined models
 include("models/Models.jl")